+++ CHANGELOG.rst
@@ -1,3 +1,12 @@
+Release 1.2.6 - Bug Fix and Enhancements
+----------------------------------------
+* Fixed issue with author ordering on initial add
+* Fixed Harvest script so that it can harvest all available articles
+* Added feature to Harvest script to allow query by date range
+* Added feature to Harvest script to optionally show progress bar
+* Added pagination to Harvest queue
+* Modified code to use the --derive flag on ldap.find_user_by_email()
+
 Release 1.2.5 - Bug Fix
 -----------------------
 * Restored dc identifiers to produce View on PubMed link